refactor launcher into new class, add rotation default, adjust layout alignment
