+++ src/launcher/CMakeLists.txt
@@ -1,7 +1,7 @@
 file(GLOB LAUNCHER_SRC_FILES CONFIGURE_DEPENDS "${CMAKE_CURRENT_LIST_DIR}/*.cpp")
 file(GLOB LAUNCHER_HDR_FILES CONFIGURE_DEPENDS "${PROJECT_SOURCE_DIR}/include/launcher/*.hpp")
 
-add_library(launcherLib SHARED ${LAUNCHER_SRC_FILES} ${LAUNCHER_HDR_FILES})
+add_library(launcherLib STATIC ${LAUNCHER_SRC_FILES} ${LAUNCHER_HDR_FILES})
 
 target_include_directories(launcherLib PRIVATE
   ${PROJECT_SOURCE_DIR}/include
@@ -9,6 +9,9 @@ target_include_directories(launcherLib PRIVATE
 
 find_package(glfw3 CONFIG REQUIRED)
 target_link_libraries(launcherLib PRIVATE glfw)
+
+find_package(imgui CONFIG REQUIRED)
+target_link_libraries(launcherLib PRIVATE imgui::imgui)
 
 find_package(CURL REQUIRED)
 target_link_libraries(launcherLib PRIVATE CURL::libcurl)
